rename pipeline package to pipelineservice
