+++ api-server/internal/pipelineservice/pipeline.go
@@ -1,4 +1,4 @@
-package pipeline
+package pipelineservice
 
 import (
 	"context"
